move regex tests to separate file
